Added safer way to consume effects, now we only expose a consumeEffects method that provides a way to consume effects from the state machine safely added JavaDoc Fixed up @StateDsl annotation for effect handler Restructured unit tests by splitting them up to the specific part of the machine what they test. Fixed namespace Upped version Updated readme
