+++ library/src/commonMain/kotlin/dev/klitsie/statemachine/StateBuilder.kt
@@ -2,6 +2,9 @@ package dev.klitsie.statemachine
 
 import kotlin.reflect.KClass
 
+/**
+ * Definition of a single state, including its transitions and side effects.
+ */
 data class StateDefinition<State : Any, CurrentState : State, Effect : Any, Event : Any>(
 	val clazz: KClass<out State>,
 	val parent: KClass<out State>?,
@@ -9,6 +12,9 @@ data class StateDefinition<State : Any, CurrentState : State, Effect : Any, Even
 	val transitions: Map<KClass<out Event>, StateTransition<State, CurrentState, Effect, Event>>,
 )
 
+/**
+ * Builder for defining a single state.
+ */
 @StateDsl
 class StateBuilder<State : Any, CurrentState : State, Effect : Any, Event : Any>(
 	private val clazz: KClass<CurrentState>,
@@ -17,12 +23,18 @@ class StateBuilder<State : Any, CurrentState : State, Effect : Any, Event : Any>
 	SideEffectBuilder<State, CurrentState, Event> by DefaultSideEffectBuilder(),
 	TransitionBuilder<State, CurrentState, Effect, Event> by DefaultTransitionBuilder() {
 
+	/**
+	 * Defines a transition for when an event of type [E] is received.
+	 */
 	inline fun <reified E : Event> onEvent(
 		noinline transition: EffectHandler<CurrentState, Effect>.(state: CurrentState, event: E) -> State,
 	) {
 		onEvent(E::class, transition)
 	}
 
+	/**
+	 * Builds the [StateDefinition].
+	 */
 	fun build(): StateDefinition<State, CurrentState, Effect, Event> {
 		return StateDefinition(
 			clazz = clazz,
