+++ library/src/commonMain/kotlin/dev/klitsie/statemachine/StateMachine.kt
@@ -5,12 +5,33 @@ import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.*
 import kotlin.reflect.KClass
 
+/**
+ * A State Machine that manages [State], handles [Event]s to transition between states,
+ * and triggers [Effect]s as side outputs.
+ *
+ * It extends [StateFlow], providing the current state as a flow.
+ *
+ * @param State The base type for all possible states in this machine.
+ * @param Effect The type for side effects triggered by transitions or states.
+ * @param Event The type for events that can be sent to the state machine.
+ */
 @OptIn(ExperimentalForInheritanceCoroutinesApi::class)
 interface StateMachine<out State : Any, Effect : Any, in Event : Any> : StateFlow<State> {
 
-	val effect: Flow<Effect>
-
+	/**
+	 * Sends an [event] to the state machine to trigger a transition.
+	 */
 	fun send(event: Event)
+
+	/**
+	 * Collects and consumes [Effect]s triggered by the state machine. Effects will be consumed if
+	 * there is at least one consumer. There is no replay cache, so effects will be consumed as they are produced.
+	 * If there is no consumer, effects will be queued until there is one. All concurrent consumers will receive
+	 * all effects.
+	 *
+	 * @param block A callback that will be invoked for each effect.
+	 */
+	suspend fun consumeEffects(block: (Effect) -> Unit)
 
 }
 
@@ -31,13 +52,11 @@ internal class DefaultStateMachine<out State : Any, Effect : Any, in Event : Any
 			.stateIn(
 				scope = scope,
 				started = started,
-				initialValue = initialState
+				initialValue = initialState,
 			)
 	}
 
-	private val _effect = Channel<Effect>()
-
-	override val effect: Flow<Effect> = _effect.receiveAsFlow()
+	private val effects = MutableStateFlow(emptyList<Effect>())
 
 	override val value: State by state::value
 	override suspend fun collect(collector: FlowCollector<State>) = state.collect(collector)
@@ -49,10 +68,32 @@ internal class DefaultStateMachine<out State : Any, Effect : Any, in Event : Any
 		}
 	}
 
-	override fun trigger(effect: Effect) = value.also {
-		scope.launch {
-			_effect.send(effect)
+	override suspend fun consumeEffects(block: (effect: Effect) -> Unit) {
+		// Make sure we are using the same context as the parent to resolve most concurrency issues in the horrid case
+		// consumeEffects is called off the main thread.
+		withContext(scope.coroutineContext.minusKey(Job.Key)) {
+			effects.collect { currentEffects ->
+				val firstEffect = currentEffects.firstOrNull() ?: return@collect
+
+				// Launch so we handle events for all consumers.
+				launch {
+					block(firstEffect)
+					effects.update { old ->
+						// Only drop if effects didn't change in the meantime. In this way, the first consumer will
+						// actually drop the effect while other consumers will still receive the effect
+						when (old) {
+							currentEffects -> old.drop(1)
+							else -> old
+						}
+					}
+					yield()
+				}
+			}
 		}
+	}
+
+	override fun trigger(effect: Effect) = value.also {
+		effects.update { it + effect }
 	}
 
 	private fun Flow<Event>.handleEvents(): Flow<State> {
@@ -72,7 +113,7 @@ internal class DefaultStateMachine<out State : Any, Effect : Any, in Event : Any
 	 */
 	private fun <T, R> Flow<T>.runningFold(
 		initial: () -> R,
-		operation: suspend (accumulator: R, value: T) -> R
+		operation: suspend (accumulator: R, value: T) -> R,
 	): Flow<R> {
 		return flow {
 			var accumulator: R = initial()
