+++ library/src/commonMain/kotlin/dev/klitsie/statemachine/StateMachineBuilder.kt
@@ -4,6 +4,19 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.SharingStarted
 import kotlin.reflect.KClass
 
+/**
+ * Creates a [StateMachine] with the specified configuration.
+ *
+ * @param State The base type for all states.
+ * @param Effect The type for side effects.
+ * @param Event The type for events.
+ * @param scope The [CoroutineScope] in which the state machine will run. This scope should ideally
+ * be single-threaded to avoid concurrency issues (e.g. `viewModelScope` or a scope with
+ * `Dispatchers.Main(.immediate)` or Dispatchers.Default.limitedParallelism(1)).
+ * @param initialState The state to start in.
+ * @param started When the state machine should start collecting (default: [SharingStarted.WhileSubscribed]).
+ * @param builder A DSL builder to define states, transitions, and side effects.
+ */
 inline fun <reified State : Any, Effect : Any, Event : Any> stateMachine(
 	scope: CoroutineScope,
 	initialState: State,
@@ -38,15 +51,24 @@ internal fun <State : Any, Effect : Any, Event : Any> stateMachine(
 
 }
 
+/**
+ * DSL marker for the StateMachine DSL.
+ */
 @DslMarker
 annotation class StateDsl
 
+/**
+ * Definition of a state machine, containing its own state definition and its sub-states.
+ */
 data class StateMachineDefinition<State : Any, CurrentState : State, Effect : Any, Event : Any>(
 	val self: StateDefinition<State, CurrentState, Effect, Event>,
 	val states: Map<KClass<out CurrentState>, StateDefinition<State, CurrentState, Effect, Event>>,
 	val nestedStates: Map<KClass<out CurrentState>, StateMachineDefinition<State, CurrentState, Effect, Event>>,
 )
 
+/**
+ * Builder for defining a state machine or a nested state.
+ */
 @StateDsl
 class StateMachineBuilder<State : Any, CurrentState : State, Effect : Any, Event : Any>(
 	private val clazz: KClass<out State>,
@@ -61,24 +83,36 @@ class StateMachineBuilder<State : Any, CurrentState : State, Effect : Any, Event
 	var nestedStates = emptyMap<KClass<out CurrentState>, StateMachineDefinition<State, CurrentState, Effect, Event>>()
 		private set
 
+	/**
+	 * Defines a transition for when an event of type [E] is received.
+	 */
 	inline fun <reified E : Event> onEvent(
 		noinline transition: EffectHandler<CurrentState, Effect>.(state: CurrentState, event: E) -> State,
 	) {
 		onEvent(E::class, transition)
 	}
 
+	/**
+	 * Defines a sub-state of type [S].
+	 */
 	inline fun <reified S : CurrentState> state(
 		noinline stateBuilder: StateBuilder<State, S, Effect, Event>.() -> Unit = {},
 	) {
 		state(S::class, stateBuilder)
 	}
 
+	/**
+	 * Defines a nested state of type [S] that can have its own sub-states.
+	 */
 	inline fun <reified S : CurrentState> nestedState(
 		noinline nestedStateBuilder: StateMachineBuilder<State, S, Effect, Event>.() -> Unit,
 	) {
 		nestedState(S::class, nestedStateBuilder)
 	}
 
+	/**
+	 * Defines a sub-state of type [S].
+	 */
 	fun <S : CurrentState> state(
 		stateClass: KClass<S>,
 		stateBuilder: StateBuilder<State, S, Effect, Event>.() -> Unit,
@@ -90,6 +124,9 @@ class StateMachineBuilder<State : Any, CurrentState : State, Effect : Any, Event
 			.plus(stateClass to stateHolder)
 	}
 
+	/**
+	 * Defines a nested state of type [S].
+	 */
 	fun <S : CurrentState> nestedState(
 		stateClass: KClass<out S>,
 		stateBuilder: StateMachineBuilder<State, S, Effect, Event>.() -> Unit,
@@ -101,6 +138,9 @@ class StateMachineBuilder<State : Any, CurrentState : State, Effect : Any, Event
 		this.nestedStates += nestedState.nestedStates.plus(stateClass to nestedState)
 	}
 
+	/**
+	 * Builds the [StateMachineDefinition].
+	 */
 	fun build(): StateMachineDefinition<State, CurrentState, Effect, Event> {
 		return StateMachineDefinition(
 			self = StateDefinition(
