+++ library/src/commonMain/kotlin/dev/klitsie/statemachine/TransitionBuilder.kt
@@ -2,18 +2,30 @@ package dev.klitsie.statemachine
 
 import kotlin.reflect.KClass
 
+/**
+ * Represents a transition from one state to another triggered by an event.
+ */
 data class StateTransition<State : Any, CurrentState : State, Effect : Any, Event : Any>(
 	val transition: EffectHandler<CurrentState, Effect>.(CurrentState, Event) -> State,
 )
 
+/**
+ * Builder for defining transitions between states.
+ */
 @StateDsl
 interface TransitionBuilder<State : Any, CurrentState : State, Effect : Any, Event : Any> {
 
+	/**
+	 * Defines a transition for the given [eventClass].
+	 */
 	fun <E : Event> onEvent(
 		eventClass: KClass<E>,
-		transition: EffectHandler<CurrentState, Effect>.(CurrentState, E) -> State
+		transition: EffectHandler<CurrentState, Effect>.(CurrentState, E) -> State,
 	)
 
+	/**
+	 * Builds the map of defined transitions.
+	 */
 	fun buildTransitions(): Map<KClass<out Event>, StateTransition<State, CurrentState, Effect, Event>>
 
 }
